+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Current:
 - disable_decays flag in advanced options fixed
 - threshold energy not used in n_mu, n_e
 - new generic function 'n_particles' for arbitrary particle types
+- new config option dedx_material
 
 Version 1.0.8:
 - Fixed a Python3 compatibility issue in density profiles